+++ src/compute-worker.js
@@ -129,7 +129,7 @@ port.on('message', /** @param {import('./compute.js').ComputeTask} msg */ msg =>
       break
     }
     case 'activity': {
-      const { nongcDoc: nongcDocBin, contentmapBin, from, to, by, contentIds: contentIdsBin, withCustomAttributions, includeCustomAttributions, includeDelta, limit, reverse, group } = msg
+      const { nongcDoc: nongcDocBin, contentmapBin, from, to, by, contentIds: contentIdsBin, withCustomAttributions, includeCustomAttributions, includeDelta, limit, reverse, group, groupMaxGap, groupMaxDuration } = msg
       const contentmap = Y.decodeContentMap(contentmapBin)
       const contentIds = contentIdsBin && Y.decodeContentIds(contentIdsBin)
       const filteredAttributions = filterContentMap(contentmap, from, to, by || undefined, contentIds, withCustomAttributions)
@@ -180,11 +180,11 @@ port.on('message', /** @param {import('./compute.js').ComputeTask} msg */ msg =>
       activity.sort((a, b) => a.from - b.from)
       /** @type {Array<{ from: number, to: number, by: string?, delta?: any, customAttributions: Array<{k:string,v:string}>|null }>} */
       const activityResult = []
-      const groupDistance = group ? 1000 : 1
+      const groupDistance = group ? groupMaxGap : 1
       /** @type {{ from: number, to: number, by: string?, customAttributions: Array<{k:string,v:string}>|null }|null} */
       let lastActivity = null
       activity.forEach(act => {
-        if (lastActivity != null && lastActivity.by === act.by && act.from - lastActivity.to < groupDistance) {
+        if (lastActivity != null && lastActivity.by === act.by && act.from - lastActivity.to < groupDistance && act.to - lastActivity.from < groupMaxDuration) {
           lastActivity.to = act.to
           lastActivity.customAttributions?.push(...(act?.customAttributions || []))
         } else {