+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## [0.4.0]
+## [0.3.1]
 
 ### New Features